requirements.txt
pymongo<4.9
fastapi==0.116.1
pydantic==2.11.7
motor==3.3.2
PyJWT==2.10.1
faiss-cpu==1.11.0
python-jose[cryptography]==3.5.0
opensearch-py==2.3.2
openai==1.98.0
slowapi==0.1.9
redis==5.0.3
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
click==8.2.1
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.47.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
pydantic-settings==2.6.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
psutil==6.0.0
httpx>=0.28.0,<0.29.0
python-multipart==0.0.20